feat: get course by id
