+++ app.js
@@ -1,6 +1,6 @@
 const express = require("express");
 const bodyParser = require("body-parser");
-const conn = require("./connection");
+const courseRouter = require("./routes/courseRouter")
 const app = express();
 const dotenv = require("dotenv");
 dotenv.config();
@@ -8,5 +8,11 @@ const PORT = process.env.PORT || 5000;
 // set body parser
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: false }));
+//set route
+
+//course
+app.use('/api/course', courseRouter);
+
+
 // buat server nya
 app.listen(PORT, () => console.log(`Server running at port: ${PORT}`));
