+++ controllers/courseController.js
@@ -12,15 +12,22 @@ exports.createData = (req, res) => {
   insertCourse(res, querySql, data);
 };
 
-exports.readData = (req, res) => {
+exports.readAllData = (req, res) => {
   const querySql = "SELECT * FROM course";
+
+  getAllCourse(res, querySql);
+};
+
+exports.readDataById = (req, res) => {
+  const id = req.params.id;
+  const querySql = `SELECT * FROM course WHERE id = ${id}`;
 
   getAllCourse(res, querySql);
 };
 
 exports.updateData = (req, res) => {
   const data = { ...req.body };
-  const querySearch = "SELECT * FROM course where id";
+  const querySearch = "SELECT * FROM course WHERE id = ?";
   const queryUpdate = "UPDATE course SET? WHERE id = ?";
 
   updateCourse(res, querySearch, queryUpdate, req.params.id, data);
@@ -28,7 +35,7 @@ exports.updateData = (req, res) => {
 
 exports.deleteData = (req, res) => {
   const data = { ...req.body };
-  const querySearch = "SELECT * FROM course where id";
+  const querySearch = "SELECT * FROM course WHER id=?";
   const queryDelete = "DELETE FROM course WHERE id = ?";
 
   deleteCourse(res, querySearch, queryDelete, req.params.id);
